fixed error on x86_64 which was caused by bad size_t types
git-svn-id: svn://svn.code.sf.net/p/cmpforopenssl/code@10 8c94e50a-443e-0410-be32-d2e341db1201

+++ src/openssl-SNAP-20071105/crypto/cmp/cmp_lib.c
@@ -727,7 +727,7 @@ ASN1_BIT_STRING *CMP_protection_new(CMP_PKIMESSAGE *pkimessage,
 	CRMF_PBMPARAMETER *pbm=NULL;
 
 	size_t protPartDerLen;
-	size_t macLen;
+	unsigned int macLen;
 	size_t maxMacLen;
 	unsigned char *protPartDer=NULL;
 	unsigned char *mac=NULL;
@@ -806,7 +806,7 @@ err:
 	*/
 int CMP_CERTSTATUS_set_certHash( CMP_CERTSTATUS *certStatus, const X509 *cert) {
 	ASN1_OCTET_STRING *certHash=NULL;
-	size_t hashLen;
+	unsigned int hashLen;
 	unsigned char hash[EVP_MAX_MD_SIZE];
 
 	if (!certStatus) goto err;
